requirements.txt
requests==2.32.2
python-dotenv==1.0.1
python-telegram-bot==13.15


